requirements.txt
pytest
playwright
pytest-playwright
allure-pytest
openpyxl